rm: move `normalize_path` to uucore and add tests
Since we don't run uucore tests by default I added them to our ci and readme
