+++ src/uu/rm/Cargo.toml
@@ -18,9 +18,10 @@ path = "src/rm.rs"
 clap = "2.33"
 walkdir = "2.2"
 remove_dir_all = "0.5.1"
-uucore = { version=">=0.0.7", package="uucore", path="../../uucore" }
+uucore = { version=">=0.0.7", package="uucore", path="../../uucore", features=["fs"] }
 uucore_procs = { version=">=0.0.5", package="uucore_procs", path="../../uucore_procs" }
 
 [[bin]]
 name = "rm"
 path = "src/main.rs"
+
